tmf: Fix series width in XYPresentationProvider
The WIDTH property should be stored as a Float, otherwise it will fail
to read properly using StyleManager.getFactorStyle().

[Fixed] Fix series width in XYPresentationProvider

Reviewed-on: https://git.eclipse.org/r/163935
Tested-by: Trace Compass Bot <tracecompass-bot@eclipse.org>
Tested-by: Genevieve Bastien <gbastien+lttng@versatic.net>

+++ tmf/org.eclipse.tracecompass.tmf.core/src/org/eclipse/tracecompass/tmf/core/presentation/XYPresentationProvider.java
@@ -114,7 +114,7 @@ public class XYPresentationProvider implements IXYPresentationProvider {
         builder.put(StyleProperties.SERIES_TYPE, seriesType);
         builder.put(StyleProperties.SERIES_STYLE, generateStyle(seriesType));
         builder.put(StyleProperties.COLOR, X11ColorUtils.toHexColor(color.getRed(), color.getGreen(), color.getBlue()));
-        builder.put(StyleProperties.WIDTH, width);
+        builder.put(StyleProperties.WIDTH, (float) width);
         return new OutputElementStyle(null, builder.build());
     }
 
@@ -125,7 +125,7 @@ public class XYPresentationProvider implements IXYPresentationProvider {
         builder.put(StyleProperties.SERIES_TYPE, seriesType);
         builder.put(StyleProperties.SERIES_STYLE, StyleProperties.SeriesStyle.NONE);
         builder.put(StyleProperties.COLOR, X11ColorUtils.toHexColor(color.getRed(), color.getGreen(), color.getBlue()));
-        builder.put(StyleProperties.WIDTH, width);
+        builder.put(StyleProperties.WIDTH, (float) width);
         builder.put(StyleProperties.SYMBOL_TYPE, generateTickStyle(seriesType));
         return new OutputElementStyle(null, builder.build());
     }
